Lab1 + change to long

+++ src/io/github/monstruos/lab2/CypherLib.java
@@ -3,17 +3,17 @@ package io.github.monstruos.lab2;
 import io.github.monstruos.lab1.CryptoLib;
 
 public class CypherLib {
-    public static Integer ShamirCyph(Integer m, Integer p, Integer ca,
-                                     Integer da, Integer cb, Integer db) {
+    public static long ShamirCyph(long m, long p, long ca,
+                                  long da, long cb, long db) {
         if ((ca * da) % (p - 1) != 1)
             return -1;
 
         if ((cb * db) % (p - 1) != 1)
             return -1;
 
-        Integer x1 = CryptoLib.powMod(m, ca, p);
-        Integer x2 = CryptoLib.powMod(x1, cb, p);
-        Integer x3 = CryptoLib.powMod(x2, da, p);
+        long x1 = CryptoLib.powMod(m, ca, p);
+        long x2 = CryptoLib.powMod(x1, cb, p);
+        long x3 = CryptoLib.powMod(x2, da, p);
 
         return CryptoLib.powMod(x3, db, p);
     }
@@ -25,7 +25,7 @@ public class CypherLib {
         Integer da = 19;
         Integer cb = 5;
         Integer db = 9;
-        Integer shamir = ShamirCyph(m, p, ca, da, cb, db);
+        long shamir = ShamirCyph(m, p, ca, da, cb, db);
         System.out.println("Shamir(" + m + ", " + p + ") = " + shamir);
     }
 }
